Reviews UI: dividers, larger summary; half-star ratings end-to-end
- Replace bordered cards with full-width ContentDivider separators; hero score ~2×, stars ≥1.5×.
- Remove Discussed filter; submit and decode ratings as Double for GraphQL Float.
- Order feedback sends half-snapped Double; chat user_review payload parses float ratings.

Made-with: Cursor

+++ Prelura-swift/Models/UserReview.swift
@@ -3,7 +3,8 @@ import Foundation
 /// A single review for a user (matches Flutter UserReviewModel / backend ReviewUserType).
 struct UserReview: Identifiable {
     let id: String
-    let rating: Int
+    /// 1.0…5.0 in 0.5 steps (matches backend `Float` / GraphQL `Float`).
+    let rating: Double
     let comment: String
     /// Optional “What went well?” tags from `rateUser(highlights:)`.
     let highlights: [String]
